Remove JUnit 4 support
