--- spring-test-smart-context/src/main/resources/META-INF/services/org.junit.platform.launcher.PostDiscoveryFilter
@@ -1,1 +1,0 @@
-com.github.seregamorph.testsmartcontext.SmartDirtiesPostDiscoveryFilter
