+++ tests/pom.xml
@@ -24,7 +24,6 @@
     </properties>
 
     <modules>
-        <module>tests-maven-junit-platform-junit4-boot24</module>
         <module>tests-maven-junit-platform-jupiter-boot35</module>
         <module>tests-maven-junit-platform-jupiter-boot40</module>
         <module>tests-maven-junit-platform-jupiter-spring5</module>
