--- tests/tests-maven-junit-platform-junit4-boot24/src/test/java/com/github/seregamorph/testsmartcontext/demo/AbstractIntegrationTest.java
@@ -1,12 +1,0 @@
-package com.github.seregamorph.testsmartcontext.demo;
-
-import com.github.seregamorph.testsmartcontext.junit4.AbstractJUnit4SpringIntegrationTest;
-import org.junit.Before;
-
-public abstract class AbstractIntegrationTest extends AbstractJUnit4SpringIntegrationTest {
-
-    @Before
-    public void beforeMethod() {
-        System.out.println("Running " + getClass());
-    }
-}
