--- tests/tests-maven-junit-platform-junit4-boot24/src/test/java/com/github/seregamorph/testsmartcontext/demo/Integration1Test.java
@@ -1,20 +1,0 @@
-package com.github.seregamorph.testsmartcontext.demo;
-
-import org.junit.Test;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.web.WebAppConfiguration;
-
-@WebAppConfiguration
-@ContextConfiguration(classes = {
-    Integration1Test.Configuration.class
-})
-public class Integration1Test extends AbstractIntegrationTest {
-
-    @Test
-    public void test() {
-    }
-
-    public static class Configuration {
-
-    }
-}
