--- tests/tests-maven-junit-platform-junit4-boot24/src/test/java/com/github/seregamorph/testsmartcontext/demo/Integration2Test.java
@@ -1,18 +1,0 @@
-package com.github.seregamorph.testsmartcontext.demo;
-
-import org.junit.Test;
-import org.springframework.test.context.ContextConfiguration;
-
-@ContextConfiguration(classes = {
-    Integration2Test.Configuration.class
-})
-public class Integration2Test extends AbstractIntegrationTest {
-
-    @Test
-    public void test() {
-    }
-
-    public static class Configuration {
-
-    }
-}
