--- tests/tests-maven-junit-platform-junit4-boot24/src/test/java/com/github/seregamorph/testsmartcontext/demo/SampleIntegrationTest.java
@@ -1,52 +1,0 @@
-package com.github.seregamorph.testsmartcontext.demo;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-import com.github.seregamorph.testsmartcontext.junit4.AbstractJUnit4SpringIntegrationTest;
-import org.junit.Before;
-import org.junit.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Bean;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.TestPropertySource;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.DefaultMockMvcBuilder;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.MOCK)
-@ContextConfiguration(classes = {
-    SampleIntegrationTest.Configuration.class
-})
-@ActiveProfiles("test")
-@TestPropertySource(properties = {
-    "parameter = value"
-})
-public class SampleIntegrationTest extends AbstractJUnit4SpringIntegrationTest {
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    @Before
-    public void beforeMethod() {
-        System.out.println("Running " + getClass());
-    }
-
-    @Test
-    public void test404() throws Exception {
-        mockMvc.perform(get("/article"))
-            .andExpect(status().isNotFound());
-    }
-
-    public static class Configuration {
-
-        @Bean
-        public MockMvc mockMvc(WebApplicationContext webApplicationContext) {
-            DefaultMockMvcBuilder builder = MockMvcBuilders.webAppContextSetup(webApplicationContext);
-            return builder.build();
-        }
-    }
-}
