--- tests/tests-maven-junit-platform-junit4-boot24/src/test/java/com/github/seregamorph/testsmartcontext/demo/Unit1Test.java
@@ -1,12 +1,0 @@
-package com.github.seregamorph.testsmartcontext.demo;
-
-import com.github.seregamorph.testsmartcontext.testkit.TestEventTracker;
-import org.junit.Test;
-
-public class Unit1Test {
-
-    @Test
-    public void test() {
-        TestEventTracker.trackEvent("Running Unit1Test");
-    }
-}
