--- tests/tests-maven-junit-platform-mixed-boot34/src/test/java/com/github/seregamorph/testsmartcontext/demo/Unit1Test.java
@@ -1,16 +1,0 @@
-package com.github.seregamorph.testsmartcontext.demo;
-
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import com.github.seregamorph.testsmartcontext.testkit.TestEventTracker;
-import org.junit.Test;
-
-// JUnit 4
-public class Unit1Test {
-
-    @Test
-    public void test() {
-        TestEventTracker.trackEvent("Running " + getClass().getName() + ".test");
-        assertTrue(true);
-    }
-}
